ingestion/requirements.txt
requests==2.28.2
pandas==1.5.3
kafka-python==2.0.2
praw==7.8.1
newsapi-python==0.2.7
psycopg2-binary==2.9.9
redis==5.0.3